intermediate-projects/music_player/pyproject.toml
[tool.poetry]
name = "music-player"
version = "0.1.0"
description = ""
authors = ["sadatyussuf <tommyleespartan0@gmail.com>"]
readme = "README.md"
packages = [{include = "music_player"}]

[tool.poetry.dependencies]
python = "^3.9"
pygame = "^2.1.2"
stagger = "^1.0.0"
pillow = "^9.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
